feat: fix message receive problem and add relevant api.

+++ src/main/java/cc/allio/turbo/modules/developer/service/impl/DevBoAttributeServiceImpl.java
@@ -155,7 +155,7 @@ public class DevBoAttributeServiceImpl extends TurboTreeCrudServiceImpl<DevBoAtt
                         .setAttrType(AttributeType.FIELD)
                         .setProps(
                                 new DevAttributeProps()
-                                        .setType(FieldType.INTEGER)
+                                        .setType(FieldType.BIGINT)
                                         .setDefaulted(true)
                                         .toJson()
                         );
